changes node relocate
